requirements.txt
# requirements.txt

django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-spectacular==0.26.5